Extract view logic to ColorLegendView
